Make grittibanzli work with non-zlib deflate streams, plus minor fixes
Grittibanzli seems to make some assumptions about the deflate streams it
is handling, which are not always met in streams created by a different
compressor than zlib. This is fixed (hopefully -- tested only with 2
sample files so far) by making it emit additional length data into the
choices file in those cases. These lengths seem to be almost always 1
(stored as 0), and are therefore low-entropy. They are appended to the
end of the choices file and therefore should not break backwards
compatibility, as grittibanzli will ignore extraneous, unknown data at
the end of this file when it is read back in upon decoding.

Also, fixed an off-by-one error for the encoded length, the comment
documenting the encoded length, and the check requiring the --meta
parameter for gzip files but not for raw deflate streams.